applyRenderingSettings() from cache to enable Save (and Undo)
